Encapsulate checklist fetching in Checklist component
